third_party/WebKit/LayoutTests/inspector/console/console-context-selector-expected.txt

Tests console execution context selector.

Console context selector:
* top 
  ____myframe (data:text/html,chro…) 
  ____⚙ worker-pause.js 

Selected worker
Console context selector:
  top 
  ____myframe (data:text/html,chro…) 
* ____⚙ worker-pause.js 

Selected iframe
Console context selector:
  top 
* ____myframe (data:text/html,chro…) 
  ____⚙ worker-pause.js 

Paused in main
Console context selector:
* top 
  ____myframe (data:text/html,chro…) [disabled]
  ____⚙ worker-pause.js 

Resumed
Console context selector:
* top 
  ____myframe (data:text/html,chro…) 
  ____⚙ worker-pause.js 

Paused in worker
Console context selector:
  top 
  ____myframe (data:text/html,chro…) 
* ____⚙ worker-pause.js 

Resumed
Console context selector:
  top 
  ____myframe (data:text/html,chro…) 
* ____⚙ worker-pause.js 

Paused in iframe
Console context selector:
  top [disabled]
* ____myframe (data:text/html,chro…) 
  ____⚙ worker-pause.js 

Resumed
Console context selector:
  top 
* ____myframe (data:text/html,chro…) 
  ____⚙ worker-pause.js 

Destroyed iframe
Console context selector:
* top 
  ____⚙ worker-pause.js 
